JENKINS-51735 - move to libs folder so it doesn't get gitignored and fix references
